# frontend/docs/bff-openapi.json
{
  "definitions": {
    "airlinepb.AirlineResponse": {
      "properties": {
        "balance": {
          "type": "number"
        },
        "created_at": {
          "type": "string"
        },
        "credit_rating": {
          "type": "integer"
        },
        "equity_cached": {
          "type": "number"
        },
        "equity_cached_at": {
          "type": "string"
        },
        "iata_code": {
          "type": "string"
        },
        "icao_code": {
          "type": "string"
        },
        "id": {
          "type": "string"
        },
        "is_bankrupt": {
          "type": "boolean"
        },
        "is_public": {
          "type": "boolean"
        },
        "name": {
          "type": "string"
        },
        "owner_id": {
          "type": "string"
        },
        "reputation": {
          "type": "integer"
        },
        "safety_rating": {
          "type": "integer"
        },
        "starting_airport_id": {
          "type": "string"
        }
      },
      "type": "object"
    },
    "airlinepb.CreateAirlineResponse": {
      "properties": {
        "balance": {
          "type": "number"
        },
        "id": {
          "type": "string"
        }
      },
      "type": "object"
    },
...
    "dto.CreateAirlineRequestDTO": {
      "properties": {
        "iata_code": {
          "example": "AA",
          "type": "string"
        },
        "icao_code": {
          "example": "AAL",
          "type": "string"
        },
        "name": {
          "example": "My Airline",
          "type": "string"
        },
        "starting_airport_id": {
          "example": "11111111-1111-1111-1111-111111111111",
          "type": "string"
        }
      },
      "type": "object"
    },
    "dto.ErrorResponse": {
      "properties": {
        "error": {
          "type": "integer"
        }
      },
      "type": "object"
    },
...
  },
  "host": "localhost:4200",
  "info": {
    "contact": {
      "email": "lboris201@yandex.ru",
      "name": "Jeno7u",
      "url": "https://github.com/Jeno7u"
    },
    "title": "AirlineSim BFF API",
    "version": "1.0",
    "description": "Backend-for-frontend API. Based on backend OpenAPI and extended with BFF cache/filter query parameters."
  },
  "paths": {
...
    "/airline": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "description": "Creates an airline for the authenticated user",
        "parameters": [
          {
            "description": "Airline details (owner_id taken from token)",
            "in": "body",
            "name": "request",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dto.CreateAirlineRequestDTO"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "Airline created successfully",
            "schema": {
              "$ref": "#/definitions/airlinepb.CreateAirlineResponse"
            }
          },
          "400": {
            "description": "1 - request validation error, 2 - airport not found, 3 - user not found",
            "schema": {
              "$ref": "#/definitions/dto.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "409": {
            "description": "Airline conflict"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Create airline",
        "tags": [
          "Airline"
        ]
      }
    },
    "/airline/{id}": {
      "get": {
        "description": "Returns full airline information by airline id",
        "parameters": [
          {
            "description": "Airline ID",
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Airline information",
            "schema": {
              "$ref": "#/definitions/airlinepb.AirlineResponse"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Airline not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Get airline by id",
        "tags": [
          "Airline"
        ]
      },
      "patch": {
        "consumes": [
          "application/json"
        ],
        "description": "Update airline name and codes (iata, icao). Only owner or admin allowed.",
        "parameters": [
          {
            "description": "Airline ID",
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          },
          {
            "description": "Fields to update: name, iata_code, icao_code",
            "in": "body",
            "name": "request",
            "required": true,
            "schema": {
              "type": "object"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/airlinepb.AirlineResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/dto.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/dto.ErrorResponse"
            }
          },
          "409": {
            "description": "Conflict"
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/dto.ErrorResponse"
            }
          }
        },
        "summary": "Update airline",
        "tags": [
          "Airline"
        ]
      }
    },
    "/airline/me": {
      "get": {
        "description": "Returns full airline information for the authenticated user",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Airline information",
            "schema": {
              "$ref": "#/definitions/airlinepb.AirlineResponse"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Airline not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Get my airline",
        "tags": [
          "Airline"
        ]
      }
    },
...
  },
  "schemes": [
    "http"
  ],
  "swagger": "2.0"
}